Moved existing classes to new CliApp project
Adopted new namespaces; nothing else.
