+++ AppMotor.CliApp/Properties/LocalizableResources.resx
