+++ AppMotor.CliApp/Terminals/IErrorOutputTerminal.cs
@@ -14,22 +14,29 @@
 // limitations under the License.
 #endregion
 
-using System;
-using System.ComponentModel;
 using System.IO;
-
-using AppMotor.Core.Extensions;
 
 using JetBrains.Annotations;
 
-namespace AppMotor.Core.System
+namespace AppMotor.CliApp.Terminals
 {
     /// <summary>
-    /// Represents a terminal/console without any "fancy" functionality; i.e. it supports reading and writing but
-    /// no interaction with the terminal window - like setting the terminal's title, cursor position or obtaining
-    /// the terminal's width. For advanced features, see <see cref="ITerminalWindow"/>.
+    /// Represents the error output of a terminal.
     /// </summary>
-    public interface ITerminal : IInputTerminal, IOutputTerminal, IErrorOutputTerminal
+    /// <seealso cref="ITerminal"/>
+    public interface IErrorOutputTerminal
     {
+        /// <summary>
+        /// The standard error output stream.
+        /// </summary>
+        [PublicAPI]
+        TextWriter Error { get; }
+
+        /// <summary>
+        /// Whether <see cref="Error"/> is redirected (to a file or the input
+        /// of another process).
+        /// </summary>
+        [PublicAPI]
+        bool IsErrorRedirected { get; }
     }
 }
