+++ AppMotor.CliApp/Terminals/Terminal.cs
@@ -24,7 +24,7 @@ using AppMotor.Core.Extensions;
 
 using JetBrains.Annotations;
 
-namespace AppMotor.Core.System
+namespace AppMotor.CliApp.Terminals
 {
     /// <summary>
     /// Replacement for <see cref="Console"/> that supports <see cref="ColoredString"/>.
