+++ AppMotor.sln
@@ -17,6 +17,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		Project.Tests.props = Project.Tests.props
 	EndProjectSection
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AppMotor.CliApp", "AppMotor.CliApp\AppMotor.CliApp.csproj", "{CBB59AA0-B953-4989-8AC9-9645AB5A0447}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AppMotor.CliApp.Tests", "AppMotor.CliApp.Tests\AppMotor.CliApp.Tests.csproj", "{7468E2E3-BDE8-4C7C-A48F-23A90D2819B0}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -31,6 +35,14 @@ Global
 		{2DBAB552-26F3-4EE8-9C9D-DF67FA933D82}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2DBAB552-26F3-4EE8-9C9D-DF67FA933D82}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2DBAB552-26F3-4EE8-9C9D-DF67FA933D82}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CBB59AA0-B953-4989-8AC9-9645AB5A0447}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CBB59AA0-B953-4989-8AC9-9645AB5A0447}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CBB59AA0-B953-4989-8AC9-9645AB5A0447}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CBB59AA0-B953-4989-8AC9-9645AB5A0447}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7468E2E3-BDE8-4C7C-A48F-23A90D2819B0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7468E2E3-BDE8-4C7C-A48F-23A90D2819B0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7468E2E3-BDE8-4C7C-A48F-23A90D2819B0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7468E2E3-BDE8-4C7C-A48F-23A90D2819B0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
